Feat/ATM-209 usdt on xtz with payoff (#112)
* Fix usdt swap contract signatures

* Fix USDt on XTZ swap contract

* Returned decimal type back for tokenId

+++ Atomex.Client.Core/Blockchain/Tezos/Tzkt/TokenInfo.cs
@@ -43,7 +43,7 @@ namespace Atomex.Blockchain.Tezos.Tzkt
             var token = new Token()
             {
                 Contract = Contract.Address,
-                TokenId = int.Parse(TokenId),
+                TokenId = decimal.Parse(TokenId),
             };
 
             if (Metadata != null)
